requirements.txt
asgiref==3.9.1
asttokens==3.0.0
decorator==5.2.1
Django==5.2.6
django-autoslug==1.9.9
django-debug-toolbar==6.0.0
django-extensions==4.1
executing==2.2.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.1.7
parso==0.8.5
pexpect==4.9.0
prompt_toolkit==3.0.52
psycopg==3.2.10
psycopg-binary==3.2.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
wcwidth==0.2.13
